% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_plot.R
\name{catplot}
\alias{catplot}
\title{Make a cat plot}
\usage{
catplot(xs, ys, size = 0.1, cat = 1, catcolor = c(0, 0, 0, 1),
  linecolor = 1, type = "justcats", canvas = c(0, 1.1, 0, 1.1), ...)
}
\arguments{
\item{xs}{a vector of numbers}

\item{ys}{another vector of numbers}

\item{size}{the size of the cat (0.1 is a good starting point)}

\item{cat}{the cat model, 1 through 12}

\item{catcolor}{a modifier vector to the png matrix (try c(1,0,0,1))}

\item{linecolor}{color of plotted lines}

\item{type}{the type of plot ... justcats, or line}

\item{canvas}{the plotting area}
}
\value{
a cat plot object... to plot more cats.
}
\description{
Make a cat plot
}
\examples{
	x <- -10:10
	y <- -x^2 + 10
	purr <- catplot(xs=x, ys=y, cat=3, catcolor=c(0,1,1,1))
	cats(purr, -x, -y, cat=4, catcolor=c(1,0,1,1))
}
