% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_plot.R
\name{cats}
\alias{cats}
\title{Plot more cats!}
\usage{
cats(obj = NULL, xs, ys, size = 0.1, cat = 1, catcolor = c(0, 0, 0, 1),
  linecolor = 1, type = "justcats")
}
\arguments{
\item{obj}{a catplot object, returned from catplot}

\item{xs}{a vector of numbers}

\item{ys}{another vector of numbers}

\item{size}{the size of the cat (0.1 is a good starting point)}

\item{cat}{the cat model, 1 through 12}

\item{catcolor}{a modifier vector to the png matrix (try c(1,0,0,1))}

\item{linecolor}{color of plotted lines}

\item{type}{the type of plot ... justcats, or line}
}
\value{
a cat plot object... to plot more cats.
}
\description{
Plot more cats!
}
