% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_cat.R
\name{morecats}
\alias{morecats}
\title{Plot even more cats! In more colors and shapes!}
\usage{
morecats(obj = NULL, xs, ys, size = 0.1, cat = c(4, 5, 6),
  catcolor = list(c(0, 0, 1, 1), c(0, 1, 0, 1)), linecolor = 1,
  type = "justcats", yshift = 0, xshift = 0)
}
\arguments{
\item{obj}{a catplot object, returned from catplot}

\item{xs}{a vector of numbers}

\item{ys}{another vector of numbers}

\item{size}{the size of the cat (0.1 is a good starting point)}

\item{cat}{the cat model, 1 through 12}

\item{catcolor}{a modifier vector to the png matrix (try c(1,0,0,1))}

\item{linecolor}{color of plotted lines}

\item{type}{the type of plot ... justcats, or line}

\item{yshift}{shifts the cat up or down, within the scaled space}

\item{xshift}{shifts the cat left or right, within the scaled space.}
}
\value{
a cat plot object... to plot more cats.
}
\description{
Plot even more cats! In more colors and shapes!
}
