% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_cat.R
\name{multicat}
\alias{multicat}
\title{Multi-cat!  More cats. More colors.}
\usage{
multicat(xs, ys, size = 0.1, cat = c(1, 2), catcolor = "#000000FF",
  linecolor = 1, type = "justcats", canvas = c(0, 1.1, 0, 1.1), ...)
}
\arguments{
\item{xs}{a vector of numbers}

\item{ys}{another vector of numbers}

\item{size}{the size of the cat (0.1 is a good starting point)}

\item{cat}{a vector of cats}

\item{catcolor}{a list of modifier vectors to the png matrix (try c(1,0,0,1))}

\item{linecolor}{color of plotted lines}

\item{type}{the type of plot ... justcats, or line}

\item{canvas}{the plotting area}

\item{...}{additional parameters to pass to plot()}
}
\value{
a cat plot object... to plot more cats.
}
\description{
Multi-cat!  More cats. More colors.
}
