% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow_cats.R
\name{rainbowCats}
\alias{rainbowCats}
\title{Rainbows!  Even more colors.}
\usage{
rainbowCats(xs, ys, ptsize = 0.1, yspread = 0.1, xspread = 0.1,
  cat = 11, catshiftx = 0, catshifty = 0, spar = NA, canvas = c(-0.5,
  1.5, -1, 1.5))
}
\arguments{
\item{xs}{a vector of numbers}

\item{ys}{another vector of numbers}

\item{ptsize}{control the size.}

\item{yspread}{the vertical spread of the rainbow}

\item{xspread}{the horizontal spread of rainbow particles}

\item{cat}{what cat shall make thine rainbow? nyan is 11}

\item{catshiftx}{get your cat connected to the rainbow!}

\item{catshifty}{get your cat connected to the rainbow!}

\item{spar}{the smoothness of the rainbow.}

\item{canvas}{you are probably going to want to zoom in and out (x1,x2,y1,y2)}
}
\value{
a rainbow!
}
\description{
Rainbows!  Even more colors.
}
\examples{
x <- -10:10
y <- -x^2 + 10
rainbowCats(x, y, yspread=0.05, xspread=0.05, canvas=c(-0.5,1.5,-1,0.8))
}
