% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_plot.R
\name{morecats}
\alias{morecats}
\title{Plot more cats in more colors and shapes!}
\usage{
morecats(obj = NULL, xs, ys, size = 0.1, cat = c(4,5,6), catcolor = list(c(0,0,1,1),c(0,1,0,1)),
  linecolor = 1, type = "justcats")
}
\arguments{
\item{obj}{a catplot object, returned from catplot}

\item{xs}{a vector of numbers}

\item{ys}{another vector of numbers}

\item{size}{the size of the cat (0.1 is a good starting point)}

\item{cat}{vector of cats, 1 through 11}

\item{catcolor}{a list of modifier vectors (to the png matrix) (try c(1,0,0,1))}

\item{linecolor}{color of plotted lines}

\item{type}{the type of plot ... justcats, or line}
}
\value{

}
\description{
Plot more cats!
}
\examples{
	x <- -10:10
	y <- -x^2 + 10
	purr <- catplot(xs=x, ys=y, cat=3, catcolor=c(0,1,1,1))
	morecats(purr, -x, -y, cat=c(4,5,6), catcolor=list(c(0,0,1,1),c(0,1,0,1)))
}
