% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_plot.R
\name{multicat}
\alias{multicat}
\title{Make a cat plot with MORE cats}
\usage{
multicat(xs, ys, size = 0.1, cat = c(1,2,3), catcolor = list(c(1,1,0,1), c(0,0,0,1),
  linecolor = 1, type = "justcats", canvas = c(0, 1.1, 0, 1.1), ...)
}
\arguments{
\item{xs}{a vector of numbers}

\item{ys}{another vector of numbers}

\item{size}{the size of the cat (0.1 is a good starting point)}

\item{cat}{a vector of cat models (currently 1...11)}

\item{catcolor}{a list of modifier vectors to the png matrix (try c(1,0,0,1))}

\item{linecolor}{color of plotted lines}

\item{type}{the type of plot ... justcats, or line}

\item{canvas}{the plotting area}
}
\value{
a cat plot object... to plot more cats.
}
\description{
Make a multi-cat plot
}
\examples{
	x <- -10:10
	y <- -x^2 + 10
	purr <- multicat(xs=x, ys=y, cat=c(1,2,3), catcolor=list(c(1,1,0,1),c(0,1,1,1)))
}
